backend/finalss/requirements.txt
# ==================================================================================
# requirements.txt - 의존성 패키지
# ==================================================================================

# --- 환경 변수 ---
python-dotenv>=1.0.0

# --- Google Gemini API ---
google-genai>=0.5.0

# --- ElevenLabs TTS (선택적) ---
elevenlabs>=1.0.0

# --- 영상 처리 ---
moviepy>=1.0.3
opencv-python>=4.8.0

# --- 이미지 처리 ---
Pillow>=10.0.0
numpy>=1.24.0

# --- 설정 파일 ---
PyYAML>=6.0

# --- 오디오 처리 ---
pydub>=0.25.1

# --- 기타 ---
# ffmpeg는 시스템에 별도 설치 필요
# Windows: https://ffmpeg.org/download.html
# Mac: brew install ffmpeg
# Linux: sudo apt install ffmpeg